refactor(stories): add story template
Refactor repetetive story into template
